+++ jobbergate-api/tests/apps/test_services.py
@@ -653,6 +653,24 @@ class TestFileService:
         file_data = await dummy_file_service.get_file_content(upserted_instance)
         assert file_data == "dummy bytes content".encode()
 
+    @pytest.mark.parametrize(
+        "filename, content",
+        [
+            ("name.py", "dummy string content"),
+            ("name.yaml", "unbalanced blackets: ]["),
+            ("name.sh.jinja2", "Hello {{ name }!"),
+            ("name.sh.j2", "Hello {{ name }!"),
+        ],
+    )
+    async def test_upsert__raises_400_on_invalid_syntax(self, filename, content, dummy_file_service):
+        """
+        Test that the ``upsert()`` method raises a 400 error if the file syntax is invalid.
+        """
+        with pytest.raises(HTTPException) as exc_info:
+            await dummy_file_service.upsert(13, filename, content)
+        assert exc_info.value.status_code == 400
+        assert "did not pass the syntax check" in exc_info.value.detail
+
     async def test_upsert__raises_500_no_size_attribute(self, make_upload_file, dummy_file_service):
         """
         Test that the ``upsert()`` method raises a 413 error if the file is too large.